requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
prisma==0.11.0
python-dotenv==1.0.0
pydantic==2.4.2
psycopg2-binary==2.9.7
pgvector==0.2.4
python-multipart==0.0.6
openai==1.12.0
pydantic-settings==2.1.0 